Move Exception declarations into own file
This is a refactor; it adds no new functionality, but is part of a larger effort to move code out of (too large) `buildozer/__init__.py`.

Also made any import lists that were touched pep8-compliant in sort order.
